Kickstart cloudflare and migrate to build dir
Make `tunnel.sh` the main entrypoint script for the container and
arrange for it to start all services. Services can now inherit the
`TUNNEL_` variables to keep settings across the various components.
`tunnel.sh` will start both tunnels, if possible as per the settings and
what is installed in the image.

Migrate the build scripts into a separate directory for an improved
structure. Also, ignore some files and directories from the Docker build
context.